requirements.txt
eth-brownie
python-dotenv
pandas
black